update editUserRole, update nugets

+++ src/RightsService.Business/Commands/User/EditUserRoleCommand.cs
@@ -1,5 +1,4 @@
-﻿using System.Collections.Generic;
-using System.Linq;
+﻿using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
 using FluentValidation.Results;
@@ -11,6 +10,7 @@ using LT.DigitalOffice.Kernel.Responses;
 using LT.DigitalOffice.RightsService.Business.Commands.User.Interfaces;
 using LT.DigitalOffice.RightsService.Data.Interfaces;
 using LT.DigitalOffice.RightsService.Mappers.Db.Interfaces;
+using LT.DigitalOffice.RightsService.Models.Db;
 using LT.DigitalOffice.RightsService.Models.Dto.Requests;
 using LT.DigitalOffice.RightsService.Validation.Interfaces;
 
@@ -56,13 +56,22 @@ namespace LT.DigitalOffice.RightsService.Business.Commands.User
 
       OperationResultResponse<bool> response = new();
 
+      DbUserRole oldUser = await _repository.GetAsync(request.UserId);
+
+      if (oldUser is null && !request.RoleId.HasValue)
+      {
+        return _responseCreator.CreateFailureResponse<bool>(HttpStatusCode.BadRequest);
+      }
+
       if (!request.RoleId.HasValue)
       {
         response.Body = await _repository.RemoveAsync(request.UserId);
       }
       else
       {
-        response.Body = (await _repository.CreateAsync(_mapper.Map(request))).HasValue;
+        response.Body = oldUser is not null
+          ? await _repository.EditAsync(oldUser, _mapper.Map(request))
+          : (await _repository.CreateAsync(_mapper.Map(request))).HasValue;
       }
 
       response.Status = response.Body
